Fix a few flaky test in css-paint-api/
The problem of these tests is that we do a takeScreenShot in the
rAF, and a rAF doesn't guarantee that the animation is ready, so
it can be easily flaky given that the animations are running on
the compositor thread.

This CL fixes it so that it calls takeScreenShot when the animation
ready promise is resolved, and the makes sure that the animation
already begun.

Bug: 1133821
Change-Id: I4fae6ccb9fe992316831d4d2013e1aa32eb52c82